Available sampling constraints

The following sampling constraints are available. These constraints may be used in any resampling setting.

Standard deviation

UncertainData.SamplingConstraints.TruncateStd — Type

TruncateStd(nσ::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at the mean ± nσ (n standard deviations).

Notes

Beware when you apply the TruncateStd constraint to a (usually a numeric) population with a small value range. With nσ small, you might end up with a population mean between the actual values, so that the range [mean(pop) - nσ*std(pop), mean(pop) + nσ*std(pop)] returns nothing.

Minimum value

UncertainData.SamplingConstraints.TruncateMinimum — Type

TruncateMinimum(min::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated below at some specified minimum value.

Maximum value

UncertainData.SamplingConstraints.TruncateMaximum — Type

TruncateMaximum(max::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated above at some specified maximum value.

Value range

UncertainData.SamplingConstraints.TruncateRange — Type

TruncateRange(min::Number, max::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at some range [min, max].

Lower quantile

UncertainData.SamplingConstraints.TruncateLowerQuantile — Type

TruncateLowerQuantile(lower_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated below at some quantile.

Upper quantile

UncertainData.SamplingConstraints.TruncateUpperQuantile — Type

TruncateUpperQuantile(upper_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated above at some quantile.

Quantile range

UncertainData.SamplingConstraints.TruncateQuantiles — Type

TruncateQuantiles(lower_quantile::Float64, upper_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at some quantile quantile (lower_quantile, upper_quantile).