Sampling constraints

UncertainData.SamplingConstraints.constrain — Method

constrain(uv::AbstractUncertainValue, constraint::SamplingConstraint)

Apply the constraint and truncate the support of the distribution furnishing the uncertain value uv. Returns a constrained uncertain value.

Element-wise constraints

The following sampling constraintsa are aimed to be used element-wise on uncertain values.

UncertainData.SamplingConstraints.TruncateStd — Type

TruncateStd(nσ::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at the mean ± nσ (n standard deviations).

Notes

Beware when you apply the TruncateStd constraint to a (usually a numeric) population with a small value range. With nσ small, you might end up with a population mean between the actual values, so that the range [mean(pop) - nσ*std(pop), mean(pop) + nσ*std(pop)] returns nothing.

UncertainData.SamplingConstraints.TruncateMinimum — Type

TruncateMinimum(min::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated below at some specified minimum value.

UncertainData.SamplingConstraints.TruncateMaximum — Type

TruncateMaximum(max::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated above at some specified maximum value.

UncertainData.SamplingConstraints.TruncateRange — Type

TruncateRange(min::Number, max::Number)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at some range [min, max].

UncertainData.SamplingConstraints.TruncateLowerQuantile — Type

TruncateLowerQuantile(lower_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated below at some quantile.

UncertainData.SamplingConstraints.TruncateUpperQuantile — Type

TruncateUpperQuantile(upper_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated above at some quantile.

UncertainData.SamplingConstraints.TruncateQuantiles — Type

TruncateQuantiles(lower_quantile::Float64, upper_quantile::Float64)

A constraint indicating that the distribution furnishing an uncertain value should be truncated at some quantile quantile (lower_quantile, upper_quantile).

Examples

using UncertainData, Distributions, Plots

# Define an uncertain value furnished by a theoretical distribution
x = UncertainValue(Normal, 1, 0.5)

# Constrain the support of the furnishing distribution using various
# constraints
xc_lq = constrain(x, TruncateLowerQuantile(0.2))
xc_uq = constrain(x, TruncateUpperQuantile(0.8))
xc_q = constrain(x, TruncateQuantiles(0.2, 0.8))
xc_min = constrain(x, TruncateMinimum(0.5))
xc_max = constrain(x, TruncateMaximum(1.5))
xc_range = constrain(x, TruncateRange(0.5, 1.5))

p_lq = plot(x, label = ""); plot!(xc_lq, label = "TruncateLowerQuantile(0.2)")
p_uq = plot(x, label = ""); plot!(xc_uq, label = "TruncateLowerQuantile(0.8)")
p_q = plot(x, label = ""); plot!(xc_q, label = "TruncateQuantiles(0.2, 0.8)")
p_min = plot(x, label = ""); plot!(xc_min, label = "TruncateMinimum(0.5)")
p_max = plot(x, label = ""); plot!(xc_max, label = "TruncateMaximum(1.5)")
p_range = plot(x, label = ""); plot!(xc_range, label = "TruncateRange(0.5, 1.5)")
plot(p_min, p_max, p_range,
    p_q, p_lq, p_uq,
    size = (750, 500), legendfont = font(7), xlabel = "Value", ylabel = "Density",
    legend = :topright, fg_legend = :transparent, bg_legend = :transparent
)

Sequential constraints

Sequential constraints are used when sampling UncertainIndexDatasets or UncertainIndexValueDatasets.

UncertainData.SamplingConstraints.StrictlyIncreasing — Type

StrictlyIncreasing(algorithm::OrderedSamplingAlgorithm; 
    n::Int = 50000, lq = 0.05, uq = 0.95)

Sampling scheme indicating element-wise sampling such that the resulting values are strictly increasing in magnitude.

Increasing sequential sampling is only guaranteed when distributions have finite support. Therefore, distributions are element-wise truncated to the lower and upper quantiles lq and uq. For each distribution, this is done by drawing n values from it, then finding the quantiles for that sample, and finally truncating the distribution to the empirical quantile range.

algorithm is an instance of some OrderedSamplingAlgorithm (e.g. StartToEnd). n is the number of samples to draw when computing quantiles.

Typically used when there are known, physical constraints on the measurements. For example, geochemical measurements of sediments at different depths of a sediment core are taken at physically separate depths in the core. Thus, the order of the indices cannot be flipped, and must be strictly decreasing/increasing.

See also: StartToEnd

UncertainData.SamplingConstraints.StrictlyDecreasing — Type

StrictlyIncreasing(algorithm::OrderedSamplingAlgorithm; n::Int = 50000)

Sampling scheme indicating element-wise sampling such that the resulting values are strictly decreasing in magnitude.

Decreasing sequential sampling is only guaranteed when distributions have finite support. Therefore, distributions are element-wise truncated to the lower and upper quantiles lq and uq. For each distribution, this is done by drawing n values from it, then finding the quantiles for that sample, and finally truncating the distribution to the empirical quantile range.

algorithm is an instance of some OrderedSamplingAlgorithm (e.g. StartToEnd). n is the number of samples to draw when computing quantiles.

Typically used when there are known, physical constraints on the measurements. For example, geochemical measurements of sediments at different depths of a sediment core are taken at physically separate depths in the core. Thus, the order of the indices cannot be flipped, and must be strictly decreasing/increasing.

See also: StartToEnd

UncertainData.SamplingConstraints.StartToEnd — Type

StartToEnd <: OrderedSamplingAlgorithm

An ordered sampling algorithm indicating that values should be treated consecutively from start to finish of the dataset.

UncertainData.SamplingConstraints.sequence_exists — Function

sequence_exists(x, c::SequentialSamplingConstraint)

Does a point-by-point sequence through the uncertain dataset x exist that satisfies the criteria c?

If x is an UncertainIndexValueDataset, then check for a sequence through the indices only.

Before the check is performed, the distributions in x are truncated to the quantiles provided by c to ensure they have finite supports.

Example

# Create a set of time indices 
# We construct this is such a way that we *know* an increasing sequence exists. 
t = [UncertainValue(Normal, i, 2) for i in 1:N];
sequence_exists(t, StrictlyIncreasing(StartToEnd()))