Types of uncertain value collections

If dealing with several uncertain values, it may be useful to represent them as an uncertain dataset. This way, one may trivially, for example, compute statistics for a dataset consisting of samples with different types of uncertainties.

Uncertain dataset types

You can collect your uncertain values in the following collections:

The UncertainValueDataset type is just a wrapper for a vector of uncertain values.

The UncertainIndexDataset type behaves just as UncertainValueDataset, but has certain resampling methods such as sequential resampling associated with them.

The UncertainIndexValueDataset type allows you to be explicit that you're working with datasets where both the indices and the data values are uncertain. This may be useful when you, for example, want to draw realizations of your dataset while simultaneously enforcing sequential resampling models. One example is resampling while ensuring the draws have strictly increasing age models.

There's also a generic uncertain dataset type for when you don't care about distinguishing between indices and data values:

UncertainDataset contains uncertain indices.

Vectors of uncertain values

Vectors of uncertain values, i.e. Vector{<:AbstractUncertainvalue}, will work seamlessly for many applications, but not for all mathematical operations and statistical algorithms. For that, rather use one of the uncertain dataset types above

Collection types

Throughout the documentation you may encounter the following type union:

UncertainData.UncertainDatasets.UVAL_COLLECTION_TYPES — Type

UVAL_COLLECTION_TYPES = Union{UD, UV} where {
    UD <: AbstractUncertainValueDataset, 
    UV <: AbstractVector{T} where {
        T <: AbstractUncertainValue}}

A type union used to represent types of uncertain values.