Statistics on datasets of uncertain values

The following statistics are available for collections of uncertain values (uncertain datasets).

Statistics.mean — Method

mean(d::AbstractUncertainValueDataset, n::Int)

Computes the element-wise mean of a dataset of uncertain values. Takes the mean of an n-draw sample for each element.

Statistics.median — Method

median(d::AbstractUncertainValueDataset, n::Int)

Computes the element-wise median of a dataset of uncertain values. Takes the median of an n-draw sample for each element.

Statistics.middle — Method

middle(d::AbstractUncertainValueDataset, n::Int)

Compute the middle of n realisations of an AbstractUncertainValueDataset.

Statistics.std — Method

std(d::AbstractUncertainValueDataset, n::Int; kwargs...)

Computes the element-wise standard deviation of a dataset of uncertain values. Takes the standard deviation of an n-draw sample for each element.

Statistics.var — Method

var(d::AbstractUncertainValueDataset, n::Int; kwargs...)

Computes the element-wise sample variance of a dataset of uncertain values. Takes the sample variance of an n-draw sample for each element.

Statistics.quantile — Method

quantile(d::AbstractUncertainValueDataset, p, n::Int; kwargs...)

Compute element-wise quantile(s) pof a dataset consisting of uncertain values. Takes the quantiles of an n-draw sample for each element.

Statistics.cov — Method

cov(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int; corrected::Bool = true)

Obtain a distribution on the covariance between two collections of uncertain values.

This is done by repeating the following procedure n times:

First, draw a length-L realisation of x by drawing one random number from each uncertain value furnishing the dataset. The draws are independent, so that no element-wise dependencies (e.g. sequential correlations) that are not already present in the data are introduced in the realisation.

Second, draw a length-L realisation of y in the same manner.

Compute the covariance between the two length-L draws.

This yields n estimates of the covariance between n independent pairs of realisations of x and y. The n-member distribution of covariance estimates is returned as a vector.

If corrected is true (the default) then the sum is scaled with n - 1 for each pair of draws, whereas the sum is scaled with n if corrected is false where n = length(x).

Statistics.cor — Method

cor(x::UVAL_COLLECTION_TYPES, y::UVAL_COLLECTION_TYPES, n::Int)

Estimate a distribution on Pearson's rank correlation coefficient between two collections of uncertain values.

This is done by repeating the following procedure n times:

First, draw a length-L realisation of x by drawing one random number from each uncertain value furnishing the dataset. The draws are independent, so that no element-wise dependencies (e.g. sequential correlations) that are not already present in the data are introduced in the realisation.

Second, draw a length-L realisation of y in the same manner.

Compute Pearson's rank correlation coefficient between the two length-L draws.

This yields n estimates of Pearson's rank correlation coefficient between n independent pairs of realisations of x and y. The n-member distribution of correlation estimates is returned as a vector.