Jarque-Bera test

Regular test

HypothesisTests.JarqueBeraTest — Type

JarqueBeraTest(d::AbstractUncertainValue, n::Int = 1000) -> JarqueBeraTest

Compute the Jarque-Bera statistic to test the null hypothesis that an uncertain value is normally distributed.

Pooled test

UncertainData.UncertainStatistics.JarqueBeraTestPooled — Function

JarqueBeraTestPooled(ud::UncertainDataset, n::Int = 1000) -> JarqueBeraTest

First, draw n realisations of each uncertain value in ud and pool them together. Then, compute the Jarque-Bera statistic to test the null hypothesis that the values of the pool are normally distributed.

Element-wise test

UncertainData.UncertainStatistics.JarqueBeraTestElementWise — Function

OneSampleADTestElementWise(ud::UncertainDataset,
    n::Int = 1000) -> Vector{JarqueBeraTest}

First, draw n realisations of each uncertain value in ud, keeping one pool of values for each uncertain value.

Then, compute the Jarque-Bera statistic to test the null hypothesis that each value pool is normally distributed.