Convenience constructors

UncertainData.UncertainValues.UncertainValue — Method

UncertainValue(d::Distributions.Distribution)

Construct an uncertain value from an instance of a distribution. If a specific uncertain value type has not been implemented, the number of parameters is determined from the distribution and an instance of one of the following types is returned:

UncertainScalarTheoreticalOneParameter

UncertainScalarTheoreticalTwoParameter

UncertainScalarTheoreticalThreeParameter

Examples

UncertainValue(Normal(0, 1))
UncertainValue(Gamma(4, 5.1))
UncertainValue(Binomial, 8, 0.2)

UncertainData.UncertainValues.UncertainValue — Method

UncertainValue(d::Type{D}, empiricaldata::Vector{T}) where {D<:Distribution, T}

Constructor for empirical distributions.

Fit a distribution of type d to the data and use that as the representation of the empirical distribution. Calls Distributions.fit behind the scenes.

Arguments

empiricaldata: The data for which to fit the distribution.

distribution: A valid univariate distribution from Distributions.jl.

UncertainData.UncertainValues.UncertainValue — Method

UncertainValue(m::Measurement) → UncertainScalarNormallyDistributed

Convert a Measurement instance to an uncertain value compatible with UncertainData.jl.

Measurement instances from Measurements.jl[1] are treated as normal distributions with known means. Once the conversion is done, the functionality provided by Measurements.jl, such as exact error propagation, is lost.

UncertainValue(values::Vector, probs::Union{Vector, AbstractWeights})

Construct a population whose members are given by values and whose sampling probabilities are given by probs. The elements of values can be either numeric or uncertain values of any type.

UncertainValue(data::Vector{T};
    kernel::Type{D} = Normal,
    npoints::Int=2048) where {D <: Distributions.Distribution, T}

Construct an uncertain value by a kernel density estimate to data.

Fast Fourier transforms are used in the kernel density estimation, so the number of points should be a power of 2 (default = 2048).

UncertainValue(kerneldensity::Type{K}, data::Vector{T};
    kernel::Type{D} = Normal,
    npoints::Int=2048) where {K <: UnivariateKDE, D <: Distribution, T}

Construct an uncertain value by a kernel density estimate to data.

Fast Fourier transforms are used in the kernel density estimation, so the number of points should be a power of 2 (default = 2048).

UncertainValue(d::Type{D}, empiricaldata::Vector{T}) where {D<:Distribution, T}

Constructor for empirical distributions.

Fit a distribution of type d to the data and use that as the representation of the empirical distribution. Calls Distributions.fit behind the scenes.

Arguments

empiricaldata: The data for which to fit the distribution.

distribution: A valid univariate distribution from Distributions.jl.

UncertainData.UncertainValues.UncertainValue — Method

UncertainValue(values::Vector, probs::Union{Vector, AbstractWeights})

Construct a population whose members are given by values and whose sampling probabilities are given by probs. The elements of values can be either numeric or uncertain values of any type.

UncertainData.UncertainValues.UncertainValue — Method

UncertainValue(x::T) where T <: Real

Create a CertainValue instance from a scalar with no uncertainty.