Convenience constructors

The following convenience constructors are used to defined uncertain values.

UncertainData.UncertainValues.UncertainValue — Function

UncertainValue(d::Distribution)
UncertainValue(d::Type{Normal}, μ, σ) → UncertainScalarNormallyDistributed
UncertainValue(d::Type{Uniform}, lower, upper) → UncertainScalarUniformlyDistributed
UncertainValue(d::Type{Beta}, α, β) → UncertainScalarBetaDistributed
UncertainValue(d::Type{BetaPrime}, α, β) → UncertainScalarBetaPrimeDistributed
UncertainValue(d::Type{Gamma}, α, θ) → UncertainScalarGammaDistributed
UncertainValue(d::Type{Frechet}, α, θ) → UncertainScalarFrechetDistributed
UncertainValue(d::Type{Binomial, n, p) → UncertainScalarBinomialDistributed
UncertainValue(d::Type{BetaBinomial, n, α, β) → UncertainScalarBetaBinomialDistributed

Construct an uncertain value represented by a (possibly truncated) theoretical distribution d.

UncertainValue(d::Type{<:Distribution}, x::AbstractVector) → UncertainScalarTheoreticalFit

Construct an uncertain value by fitting a distribution of type d to an empirical sample x, and use that fitted distribution as the representation of x.

See also: UncertainScalarTheoreticalFit

UncertainValue(x::AbstractVector; 
    kernel::Type{<:Distribution} = Normal, npoints::Int = 2048) → UncertainScalarKDE

Construct an uncertain value by estimating the underlying distribution to the empirical sample x using the kernel density estimation (KDE), then using the resulting KDE-distribution as the representation of x. Fast Fourier transforms are used in the kernel density estimation, so the number of points should be a power of 2 (default = 2048).

See also: UncertainScalarKDE

UncertainValue(pop::Vector, probs::Union{Vector, AbstractWeights}) → UncertainScalarPopulation

Construct an uncertain value from a population pop, whose sampling probabilities (prior beliefs) are probs. The population pop can contain any type of uncertain value. Scalars in pop are converted to CertainScalars.

See also: UncertainScalarPopulation

UncertainValue(x::T) where {T <: Real} → CertainScalar

Create a CertainScalar instance from a scalar with no uncertainty.

See also: CertainScalar

UncertainValue(m::Measurement) → UncertainScalarNormallyDistributed

Convert a Measurement instance to an uncertain value compatible with UncertainData.jl.

Measurement instances from Measurements.jl[1] are treated as normal distributions with known means. Once the conversion is done, the functionality provided by Measurements.jl, such as exact error propagation, is lost.

Examples

Theoretical distributions with known parameters

Measurements are often given as a mean and an associated standard deviation. Such measurements can be directly represented by the parameters of the distribution.

Assume a data point has a normally distributed uncertainty, with a mean value of 2.2 and standard deviation of 4.0. We use the following notation to represent that value.

using UncertainData, Distributions
UncertainValue(Normal(2.2, 4.0))
UncertainValue(Normal, 2.2, 4.0) # alternative constructor

Other distributions, as well as truncated distributions, also work.

using UncertainData
UncertainValue(Uniform, -5.0, 5.0)
UncertainValue(Gamma, 3.0, 1.2)

lo, hi = 0.5, 3.5 # truncation limits
UncertainValue(Truncated(Gamma(4, 5.1), lo, hi))

Theoretical distributions with parameters estimated from empirical data

In some cases, it might be convenient to represent an empirical sample by a porobability distribution whose parameters are estimated from the sample. Here, we simulate a real dataset by generating a small sample from a normal distribution, then fit a normal distribution to it.

using UncertainData, Distributions
s = rand(Normal(0, 1), 100)

# Represent the sample `s` by a normal distribution with estimated parameters
x = UncertainValue(Normal, s)

Distributions estimated using the kernel density approach

For empirical data with non-trivial underlying distributions, one may use kernel density estimation to fit a distribution to the empirical sample.

Below, we simulate a multimodal empirial sample, and represent that sample by a kernel density estimated distribution.

using UncertainData, Distributions
M = MixtureModel(Normal[
          Normal(-2.0, 1.2),
          Normal(0.0, 1.0),
          Normal(3.0, 2.5)], [0.2, 0.5, 0.3])
# This is our sample
s = rand(M, 40000)

# `x` is now a kernel density estimated distribution that represents the sample `s`
x = UncertainValue(s) # or UncertainValue(UnivariateKDE, s) to be explicit

Populations (discrete sets of values with associated weights)

Sometimes, numerous measurements of the same phenomenon might be available. In such cases, a population may be used to simultaneously represent all data available. Weights representing prior beliefs can be added (set weights equal if all points are equiprobable).

Below, we assume x1 and x2 were measured with sophisticated devices, giving both a mean and standard deviation. x3, on the other hand, was measured with a primitive device, giving only a mean value. Hence our trust in x3 is lower than for x1 and x2. The following

x1 = UncertainValue(Normal, 0.1, 0.5)
x2 = UncertainValue(Gamma, 1.2, 3.1)
x3 = UncertainValue(0.1)
pop = [x1, x2, x3] # the population
wts = [0.45, 0.45, 0.1] # weights; `x1` and `x2` are equiprobable, and more probable than `x3`.
UncertainValue(pop, wts)

Values without uncertainties

Numerical values without associated uncertainties must be converted before mixing with uncertain values.

x = UncertainValue(2.0)

Compatibility with Measurements.jl

Measurements from Measurements.jl are assumed to be normally distributed and errors are propagated using linear error propagation theory. In this package, resampling is used to propagate errors. Thus, Measurements must be converted to normal distributions to be used in conjuction with other uncertain values in this package.

using UncertainData, Measurements
m = measurement(value, uncertainty)
x = UncertainValue(m) # now compatible with UncertainData.jl, but drops support for exact error propagation