Increasing/decreasing

The following constraints may be used to impose sequential constraints when sampling a collection of uncertain values element-wise.

StrictlyIncreasing

UncertainData.SamplingConstraints.StrictlyIncreasing — Type

StrictlyIncreasing

A sampling constraint indicating element-wise sampling of the uncertain values in a dataset, such that the values of the draw are strictly increasing in magnitude.

Typically used when there are known, physical constraints on the measurements. For example, geochemical measurements of sediments at different depths of a sediment core are taken at physically separate depths in the core. Thus, the order of the indices cannot be flipped, and must be strictly decreasing/increasing.

StrictlyDecreasing

UncertainData.SamplingConstraints.StrictlyDecreasing — Type

StrictlyDecreasing

A sampling constraint indicating element-wise sampling of the uncertain values in a dataset, such that the values of the draw are strictly decreasing in magnitude.

Typically used when there are known, physical constraints on the measurements. For example, geochemical measurements of sediments at different depths of a sediment core are taken at physically separate depths in the core. Thus, the order of the indices cannot be flipped, and must be strictly decreasing/increasing.

Checking for increasing/decreasing sequences

There are a few built-in functions to check if your dataset allows the application of certain sequential sampling constraints. These functions will check whether a valid sequence through your collection of uncertain values exists, so that you can know beforehand whether a particular resampling scheme is possible to apply to your data.

Strictly increasing

UncertainData.SamplingConstraints.strictly_increasing_sequence_exists — Function

strictly_increasing_sequence_exists(udata::AbstractUncertainValueDataset; 
    quantiles = [0.0001, 0.9999])

Does a path through the dataset exist? I.e, check that a strictly increasing sequence can be found after first constraining each distribution to the provided quantile range (this is necessary because some distributions may have infinite support).

Strictly decreasing

UncertainData.SamplingConstraints.strictly_decreasing_sequence_exists — Function

strictly_decreasing_sequence_exists(udata::AbstractUncertainValueDataset;
    quantiles = [0.0001, 0.9999])

Does a path through the dataset exist? I.e, check that a strictly decreasing sequence can be found after first constraining each distribution to the provided quantile range (this is necessary because some distributions may have infinite support).